tidy(config): move config docstring builder to its script
